+++ packages/legend-application/src/components/ActionAlert.tsx
@@ -43,7 +43,7 @@ const ActionAlertInner = observer((props: { info: ActionAlertInfo }) => {
     onClose?.();
     applicationStore.setActionAltertInfo(undefined);
   };
-  const handlEnter = (): void => onEnter?.();
+  const handleEnter = (): void => onEnter?.();
   const handleSubmit = (): void => {
     const proceedActions = actions.filter((action) => action.default);
     if (proceedActions.length) {
@@ -61,7 +61,7 @@ const ActionAlertInner = observer((props: { info: ActionAlertInfo }) => {
       open={Boolean(applicationStore.actionAlertInfo)}
       onClose={noop} // disallow closing dialog by using Esc key or clicking on the backdrop
       TransitionProps={{
-        onEnter: handlEnter,
+        onEnter: handleEnter,
       }}
     >
       <form