+++ packages/legend-extension-dsl-data-space/src/models/__tests__/TEST_DATA__DSLDataSpace_Roundtrip.ts
@@ -19,6 +19,21 @@ export const TEST_DATA__roundtrip = [
     path: 'model::testDataSpace',
     content: {
       _type: 'dataSpace',
+      stereotypes: [
+        {
+          profile: 'meta::pure::profiles::typemodifiers',
+          value: 'abstract',
+        },
+      ],
+      taggedValues: [
+        {
+          tag: {
+            profile: 'meta::pure::profiles::doc',
+            value: 'doc',
+          },
+          value: 'something',
+        },
+      ],
       artifactId: 'legend-query-test',
       groupId: 'org.finos.legend.test',
       mapping: 'dummyMapping',