+++ packages/legend-query/CHANGELOG.md
@@ -20,4 +20,4 @@
 
 ### Patch Changes
 
-- [#427](https://github.com/finos/legend-studio/pull/427) [`23b59b89`](https://github.com/finos/legend-studio/commit/23b59b8962c5049d1605bcb262c16cd3c012a1dd) ([@akphi](https://github.com/akphi)) - Move `@finos/legend-studio-preset-query-builder` core logic into `@finos/legend-query`, hence sever the dependency of `Legend Query` core on `@finos/legend-studio`.
+- [#427](https://github.com/finos/legend-studio/pull/427) [`23b59b89`](https://github.com/finos/legend-studio/commit/23b59b8962c5049d1605bcb262c16cd3c012a1dd) ([@akphi](https://github.com/akphi)) - Move `@finos/legend-studio-extension-query-builder` core logic into `@finos/legend-query`, hence sever the dependency of `Legend Query` core on `@finos/legend-studio`.